requirements.txt
mysqlclient==2.0.3
Django==3.1.4
djangorestframework==3.12.2
drf-yasg==1.20.0
coverage==5.3.1
libgravatar==0.2.4
django-filter==2.4.0
gunicorn==20.0.4
django-cors-headers==3.7.0
pytz~=2021.1
# django-crontab==0.7.1
requests==2.25.1
lxml==4.6.1